cli() function
