+++ package.json
@@ -15,6 +15,7 @@
     "autotest": "npm test -- --watch"
   },
   "dependencies": {
+    "bluebird": "^2.9.9",
     "chalk": "^0.5.1",
     "commander": "^2.6.0",
     "date-fns": "git://github.com/js-fns/date-fns",
