+++ src/__tests__/cli_test.js
@@ -1,0 +1,5 @@
+var cli = require('../cli');
+
+describe('cli', function() {
+  // TODO: @2015-02-16 write tests
+});
