+++ src/run_editor.js
@@ -1,7 +1,10 @@
 var childProcess = require('child_process');
 
-var runEditor = function(todo, option) {
+var runEditor = function(todo, option, cb) {
   var vimPrc = childProcess.spawn('vim', [todo.filename, '+' + todo.lineNumber], {stdio: 'inherit'});
+  vimPrc.on('close', function(code) {
+    cb(code);
+  });
 }
 
 module.exports = runEditor;
